requirements.txt
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dnspython==2.3.0
email-validator==1.3.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mobility==1.1.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
fonttools==4.39.2
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.7.1
numpy==1.24.2
packaging==23.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
six==1.16.0
SQLAlchemy==2.0.4
typing-extensions==4.3.0
urllib3==1.26.15
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
